refactor(utxo-coredao): rename from transaction to opReturn
We will make a separate transaction module later

TICKET: BTC-1578
