+++ modules/utxo-coredao/src/index.ts
@@ -1,1 +1,1 @@
-export * from './transaction';
+export * from './opReturn';
